Fix eslint & test
