+++ src/components/screens/request/index.js
@@ -1,35 +1,5 @@
-import PropTypes from 'prop-types';
-import React from 'react';
 import { connect } from 'react-redux';
-import { tokenKeys } from '../../../constants/tokens';
-import RequestBtc from './requestBtc';
-import RequestLsk from './requestLsk';
-
-const TagNameMap = {
-  LSK: RequestLsk,
-  BTC: RequestBtc,
-};
-
-const Request = ({
-  account, t, token,
-}) => {
-  const TagName = TagNameMap[token];
-  const address = account.info ? account.info[token].address : '';
-  return <TagName address={address} t={t} />;
-};
-
-Request.propTypes = {
-  address: PropTypes.string.isRequired,
-  token: PropTypes.oneOf(tokenKeys).isRequired,
-  t: PropTypes.func.isRequired,
-};
-
-Request.defaultProps = {
-  address: '',
-  token: 'LSK',
-  /* istanbul ignore next */
-  t: key => key,
-};
+import Request from './request';
 
 export default connect(
   state => ({
