+++ src/components/screens/wallet/overview/accountInfo/accountInfo.css
@@ -107,7 +107,7 @@
       display: flex;
       align-items: center;
       justify-content: center;
-      
+
       & > div {
         display: flex;
       }
